requirements.txt
aiogram>=3.13.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pillow>=10.0.0
